+++ api/src/main/java/pl/kolendateam/dadcard/skills/dto/RaceSkillsDTO.java
@@ -3,6 +3,6 @@ package pl.kolendateam.dadcard.skills.dto;
 public class RaceSkillsDTO {
 
     public String nameSkill;
-    public int skillRank;
+    public double skillRank;
     
 }
